Move object store test upload test cases to test/integration/objectstore
